Added new option: F4 compression bank 0 skip
For F4 bankswitching on bank has to be compressed. If bank 0 is chosen to
be compressed, this will not bank selected after power on. The PLD boards
always start on bank 0. At least one ROM (Triptych) requires this.
Selection this option will make sure that the first bank is never
compressed.

Signed-off-by: Sven Oliver Moll <svolli@svolli.de>